Add tests for displaying sleep sessions in the sleep log
This will also switch react-scripts with react-app-rewired for the
start, build, and test commands. The reason is because the package
msw was introduced to allow mocking server calls in unit tests.

This project was created with react-scripts@5.0.1, which uses webpack 5.
Webpack 5 no longer contains polyfills for node.js core modules by
default, but the default react configuration does not allow for easy
modification.

In order to work around this, react-app-rewired will be added and a
configuration override file will be created to resolve the dependency
issues. Likewise, all of the libraries will be introduced into the
project as development dependencies.

Relevant links
https://docs.immutable.com/docs/x/create-react-app-webpack-5-errors/